Fix price type to float

+++ inventory-service/lib/db.rb
@@ -5,7 +5,7 @@ DB = Sequel.sqlite # memory database, requires sqlite3
 DB.create_table :items do
   primary_key :id
   String :name, null: false
-  Integer :price, null: false
+  Float :price, null: false
   TrueClass :active, default: true
 end
